+++ enterprise/models.py
@@ -2,42 +2,108 @@
 """
 Database models for enterprise.
 """
-
 from __future__ import absolute_import, unicode_literals
 
-# from django.db import models
-from django.utils.encoding import python_2_unicode_compatible
+from uuid import uuid4
 
+from simple_history.models import HistoricalRecords  # likely a bug in import order checker
+from django.db import models
+from django.utils.encoding import python_2_unicode_compatible
+from django.utils.translation import ugettext_lazy as _
 from model_utils.models import TimeStampedModel
+
+
+class EnterpriseCustomerManager(models.Manager):
+    """
+    Model manager for Enterprise Customer model.
+
+    Filters out inactive Enterprise Customers, otherwise works the same as default model manager.
+    """
+
+    # This manager filters out some records, hence according to the Django docs it must not be used
+    # for related field access. Although False is default value, it still makes sense to set it explicitly
+    # https://docs.djangoproject.com/en/1.10/topics/db/managers/#base-managers
+    use_for_related_fields = False
+
+    def get_queryset(self):
+        """
+        Return a new QuerySet object. Filters out inactive Enterprise Customers.
+        """
+        return super(EnterpriseCustomerManager, self).get_queryset().filter(active=True)
 
 
 @python_2_unicode_compatible
 class EnterpriseCustomer(TimeStampedModel):
     """
-    TODO: replace with a brief description of the model.
+    Enterprise Customer is an organization or a group of people that "consumes" courses.
+
+    Users associated with an Enterprise Customer take courses on the edX platform.
+
+    Enterprise Customer might be providing certain benefits to their members, like discounts to paid course
+    enrollments, and also might request (or require) sharing learner results with them.
+
+    Fields:
+        uuid (UUIDField, PRIMARY KEY): Enterprise Customer code - used to reference this Enterprise Customer in
+            `                                other parts of the system (SSO, ecommerce, analytics etc.)
+        name (CharField): Enterprise Customer name
+        active (BooleanField): used to mark inactive Enterprise Customers - implements "soft delete" pattern
     """
 
-    # TODO: add field definitions
+    class Meta(object):
+        verbose_name = _("Enterprise Customer")
+        verbose_name_plural = _("Enterprise Customers")
+
+    objects = models.Manager()
+    active_customers = EnterpriseCustomerManager()
+
+    uuid = models.UUIDField(primary_key=True, default=uuid4, editable=False)
+    name = models.CharField(max_length=255, blank=False, null=False, help_text=_("Enterprise Customer name."))
+    active = models.BooleanField(default=True)
+    history = HistoricalRecords()
 
     def __str__(self):
         """
-        Get a string representation of this model instance.
+        Return human-readable string representation.
         """
-        # TODO: return a string appropriate for the data fields
-        return '<EnterpriseCustomer, ID: {}>'.format(self.id)
+        return "<EnterpriseCustomer {code}: {name}>".format(code=self.uuid, name=self.name)
+
+    def __repr__(self):
+        """
+        Return uniquely identifying string representation.
+        """
+        return self.__str__()
 
 
 @python_2_unicode_compatible
-class EnterpriseCustomerHistory(TimeStampedModel):
+class EnterpriseCustomerUser(TimeStampedModel):
     """
-    TODO: replace with a brief description of the model.
+    Model that keeps track of user - enterprise customer affinity.
+
+    Fields:
+        enterprise_customer (ForeignKey[EnterpriseCustomer]): enterprise customer
+        user_id (IntegerField): user identifier
     """
 
-    # TODO: add field definitions
+    enterprise_customer = models.ForeignKey(EnterpriseCustomer, blank=False, null=False)
+    user_id = models.PositiveIntegerField(null=False, blank=False)
+
+    class Meta(object):
+        verbose_name = _("Enterprise Customer User")
+        verbose_name_plural = _("Enterprise Customer Users")
+        unique_together = (("enterprise_customer", "user_id"),)
 
     def __str__(self):
         """
-        Get a string representation of this model instance.
+        Return human-readable string representation.
         """
-        # TODO: return a string appropriate for the data fields
-        return '<EnterpriseCustomerHistory, ID: {}>'.format(self.id)
+        return "<EnterpriseCustomerUser {ID}: {enterprise_name} - {user_id}>".format(
+            ID=self.id,
+            enterprise_name=self.enterprise_customer.name,
+            user_id=self.user_id,
+        )
+
+    def __repr__(self):
+        """
+        Return uniquely identifying string representation.
+        """
+        return self.__str__()