+++ test_utils/factories.py
@@ -1,0 +1,52 @@
+"""
+Factoryboy factories.
+"""
+from __future__ import absolute_import, unicode_literals
+
+from uuid import UUID
+
+import factory
+from faker import Factory as FakerFactory
+
+from enterprise.models import EnterpriseCustomer, EnterpriseCustomerUser
+
+FAKER = FakerFactory.create()
+
+
+class EnterpriseCustomerFactory(factory.django.DjangoModelFactory):
+    """
+    EnterpriseCustomer factory.
+
+    Creates an instance of EnterpriseCustomer with minimal boilerplate - uses this class' attributes as default
+    parameters for EnterpriseCustomer constructor.
+    """
+
+    class Meta(object):
+        """
+        Meta for EnterpriseCustomerFactory.
+        """
+
+        model = EnterpriseCustomer
+
+    uuid = factory.LazyAttribute(lambda x: UUID(FAKER.uuid4()))
+    name = factory.LazyAttribute(lambda x: FAKER.company())
+    active = True
+
+
+class EnterpriseCustomerUserFactory(factory.django.DjangoModelFactory):
+    """
+    EnterpriseCustomer factory.
+
+    Creates an instance of EnterpriseCustomerUser with minimal boilerplate - uses this class' attributes as default
+    parameters for EnterpriseCustomerUser constructor.
+    """
+
+    class Meta(object):
+        """
+        Meta for EnterpriseCustomerFactory.
+        """
+
+        model = EnterpriseCustomerUser
+
+    enterprise_customer = factory.SubFactory(EnterpriseCustomerFactory)
+    user_id = factory.LazyAttribute(lambda x: FAKER.pyint())