+++ CHANGELOG.rst
@@ -15,9 +15,10 @@ Unreleased
 ~~~~~~~~~~
 
 [0.1.0] - 2016-10-13
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~
 
 Added
 _____
 
 * First release on PyPI.
+* Models and Django admin integration